github_issue_lister_agent_vllm/pyproject.toml
[project]
name = "github-issue-lister"
version = "0.1.0"
description = "An ADK-compatible agent to list GitHub issues."
authors = [{ name = "Gemini", email = "gemini@google.com" }]
dependencies = [
    "PyGithub==2.6.1",
    "google-adk==1.6.1",
    "mcp==1.12.0",
    "cloudpickle==3.1.1",
]

[project.scripts]
github-issue-lister = "main:main"

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"
